refactor: extracted reducer function to a new file
