+++ src/context/auth-context.js
@@ -1,4 +1,5 @@
 import { createContext, useContext, useReducer, useEffect } from "react";
+import { authReducer, initialState } from "../reducer/auth-reducer";
 
 const AuthContext = createContext();
 
@@ -8,69 +9,6 @@ const AuthProvider = ({ children }) => {
          ? authDispatch({ type: "TOGGLE_LOGIN" })
          : "";
    }, []);
-
-   const authReducer = (authState, { type, payload }) => {
-      switch (type) {
-         case "TOGGLE_LOGIN":
-            return {
-               ...authState,
-               isLoggedIn: !authState.isLoggedIn,
-            };
-         case "SET_EMAIL":
-            return {
-               ...authState,
-               email: payload,
-            };
-         case "SET_PW":
-            return {
-               ...authState,
-               password: payload,
-            };
-         case "SET_NAME":
-            return {
-               ...authState,
-               name: payload,
-            };
-         case "SET_ERROR":
-            return {
-               ...authState,
-               error: payload,
-            };
-         case "SET_GUEST_CREDENTIALS":
-            return {
-               ...authState,
-               email: payload[0],
-               password: payload[1],
-               error: "",
-            };
-         case "SET_USER_CREDENTIALS":
-            return {
-               ...authState,
-               name: payload.name,
-               email: payload.email,
-               password: payload.password ?? "",
-               isLoggedIn: !authState.isLoggedIn,
-            };
-         case "CLEAR_FIELDS":
-            return {
-               ...authState,
-               name: "",
-               email: "",
-               password: "",
-               error: "",
-            };
-         default:
-            authState;
-      }
-   };
-
-   const initialState = {
-      isLoggedIn: false,
-      name: "",
-      email: "",
-      password: "",
-      error: "",
-   };
 
    const [authState, authDispatch] = useReducer(authReducer, initialState);
